Add Worekers and Queue Fixes #6
