+++ app/models/__init__.py
@@ -1,3 +1,3 @@
-from .models import User, Project, Scene, ProjectStatus, ChatMessage, Media, Status, MediaType, ChatRole, Media, MediaType
+from .models import User, Project, Scene, ChatMessage, Media, Status, MediaType, ChatRole, Media, MediaType
 # Optionally, for explicit export
 __all__ = ["User", "Project", "Scene"]
