+++ app/models/enums.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class ProjectStatus(str, Enum):
+    queued= 'queued'
+    processing = 'processing'
+    complete = "complete"
+    error = "error"
