+++ app/services/__init__.py
@@ -1,6 +1,6 @@
-from .llm_client import generate_code
+from .llm_client import generate_code, generate_code_new
 from .manim_engine import edit_manim, generate_manim_file,create_seperate_scenes, update_scene,edit_scene, create_single_scene
 from .services import parse_gemini_response, parse_single_scene
 from .prompts import BASE_PROMPT, RE_BASE_PROMPT,SCENE_PROMPT
-from. create_manim_project import updateSceneFile, create_folder_for_project, create_manim_project, create_db_entries, apply_bolt_artifact, initialize_project, run_manim, merge_llm_response
+from. create_manim_project import updateSceneFile, create_folder_for_project, create_manim_project, create_db_entries, apply_bolt_artifact, initialize_project, merge_llm_response
 from .updated_prompts import systemPrompt, BASE_PROMPT, editSystemPrompt
